Initial tests to see what is possible.
Current design being tested is to have a REST API backend that communicates with a JS frontend.
It is unclear how to best handle authenticated endpoints with this design.
It would be ideal to use the appengine's User package but that invovles doing some strange redirects.
This approach may not work when the frontend and backend are hosted on different domains.