Changed member model to no longer have an array of event signups. Instead, now an event signup is a child of a member, and it has the same key name as the event that it belongs to.
TODO: Determine if it is a better idea to have an event signup be a
child of the event, and then the signup would have to store the ID of
the user who signed up for it.  The application would be completely
respoinsible for preventing the user from signing up for the same event
multiple times (whereas in the current model the datastore prevents the
user from signing up for the same event multiple times - at the worst
case the event signup entity would just get updated, a new one would
not be created).