Added API call to create an event via a JSON POST request. The user must be logged in as an admin to create an event.
Modified responseJSON to have extra fields that can be used to
communicate an entity and its ID.  This should probably be broken out
into its own struct in the future once the API is finalized.

The allEvents API call now returns a list of responseJSON so that the
client is aware of both the event and its ID.  This will be used to
modify the event signup process so that it will no longer be based on
an event having a static name as the key string ID.

Created constants to represent the various entites in the datastore.